Added Source code
CLTNode: a Class that defines each course as a node with relations to other nodes. (Flow Chart Driven).
CLT: a class that handles the net of nodes.
Test: a Test class where the courses are added, relations between courses established and plans suggested.   (No GUI Currently)

Works only for SWE major, since it's the only course with added relations,Although it's written as SWE plan, any major will work, just needs a change on names and relations following that major's Flow Chart.   (Put your courses instead of Mine!)